refactor: DELETE 배치 처리 및 포럼 채널 삭제 수정

+++ src/backend/community-service/src/main/java/capstone/communityservice/domain/server/service/ServerCommandService.java
@@ -280,7 +280,7 @@ public class ServerCommandService {
         }
     }
     private void validateServerProfileDelete(Server server) {
-        if(server.getProfile() != null)
+        if(server.getProfile() != null && !server.getProfile().equals("null"))
             fileUploadService.delete(server.getProfile());
     }
